Update LQR_Controller_for_Bicycle_Model.py
This is the current version of the LQR controller for a car modeled as a bicycle. It's designed to take a car slightly offset from a path, steer it onto the path, and get it to follow said path while remaining front-facing and relatively aligned with (AKA tangent to) said path at all times.